Reduce example duplication and stabilize sweep screenshots
