+++ CodeGlyphX.Examples/ExamplePaths.cs
@@ -1,0 +1,24 @@
+using System;
+using System.IO;
+
+namespace CodeGlyphX.Examples;
+
+internal static class ExamplePaths {
+    public static string ResolveSamplesDir(string relativePath) {
+        if (string.IsNullOrWhiteSpace(relativePath)) {
+            throw new ArgumentException("Path is required.", nameof(relativePath));
+        }
+
+        var dir = new DirectoryInfo(AppContext.BaseDirectory);
+        for (var i = 0; i < 10 && dir is not null; i++) {
+            var candidate = Path.Combine(dir.FullName, relativePath);
+            if (Directory.Exists(candidate)) {
+                return candidate;
+            }
+
+            dir = dir.Parent;
+        }
+
+        throw new DirectoryNotFoundException($"Could not locate sample directory '{relativePath}'.");
+    }
+}
