+++ CodeGlyphX.Examples/QrDecodeSamplesExample.cs
@@ -11,7 +11,7 @@ internal static class QrDecodeSamplesExample {
     private static readonly string[] ImageExtensions = { ".png", ".jpg", ".jpeg", ".bmp", ".gif", ".webp", ".tga" };
 
     public static void Run(string outputDir) {
-        var sampleDir = ResolveSamplesDir("Assets/DecodingSamples");
+        var sampleDir = ExamplePaths.ResolveSamplesDir("Assets/DecodingSamples");
         var files = Directory.EnumerateFiles(sampleDir)
             .Where(path => ImageExtensions.Contains(Path.GetExtension(path), StringComparer.OrdinalIgnoreCase))
             .OrderBy(path => path, StringComparer.OrdinalIgnoreCase)
@@ -61,17 +61,7 @@ internal static class QrDecodeSamplesExample {
                 lines.Add($"{name}: FAIL {info}");
                 if (debugOutputs && ImageReader.TryDecodeRgba32(bytes, out var rgba, out var width, out var height)) {
                     var baseName = Path.GetFileNameWithoutExtension(name);
-                    var opts = new QrPixelDebugOptions();
-                    QrPixelDebug.RenderToFile(rgba, width, height, width * 4, PixelFormat.Rgba32, QrPixelDebugMode.Binarized, debugDir, $"{baseName}-bin.png", opts);
-
-                    var adaptive = new QrPixelDebugOptions {
-                        AdaptiveThreshold = true,
-                        AdaptiveWindowSize = 15,
-                        AdaptiveOffset = 8
-                    };
-                    QrPixelDebug.RenderToFile(rgba, width, height, width * 4, PixelFormat.Rgba32, QrPixelDebugMode.Binarized, debugDir, $"{baseName}-bin-adaptive.png", adaptive);
-
-                    QrPixelDebug.RenderToFile(rgba, width, height, width * 4, PixelFormat.Rgba32, QrPixelDebugMode.Heatmap, debugDir, $"{baseName}-heatmap.png", opts);
+                    ExampleDebug.WriteQrDebugImages(rgba, width, height, width * 4, debugDir, baseName);
                 }
             }
         }
@@ -114,23 +104,5 @@ internal static class QrDecodeSamplesExample {
     private static bool ReadBoolEnv(string name, bool fallback) {
         var value = Environment.GetEnvironmentVariable(name);
         return bool.TryParse(value, out var parsed) ? parsed : fallback;
-    }
-
-    private static string ResolveSamplesDir(string relativePath) {
-        if (string.IsNullOrWhiteSpace(relativePath)) {
-            throw new ArgumentException("Path is required.", nameof(relativePath));
-        }
-
-        var dir = new DirectoryInfo(AppContext.BaseDirectory);
-        for (var i = 0; i < 10 && dir is not null; i++) {
-            var candidate = Path.Combine(dir.FullName, relativePath);
-            if (Directory.Exists(candidate)) {
-                return candidate;
-            }
-
-            dir = dir.Parent;
-        }
-
-        throw new DirectoryNotFoundException($"Could not locate sample directory '{relativePath}'.");
     }
 }
