+++ CodeGlyphX.Tests/QrDecodingSamplesSweepTests.cs
@@ -61,7 +61,7 @@ public sealed class QrDecodingSamplesSweepTests {
                     BudgetMilliseconds = 12000,
                     AutoCrop = true,
                     AggressiveSampling = true,
-                    StylizedSampling = true,
+                    StylizedSampling = false,
                     EnableTileScan = true,
                     TileGrid = 6
                 };
